requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
torch>=1.9.0
pytest>=6.0.0
pytest-asyncio>=0.18.0